requirements.txt
python-whois>=0.9.4
requests>=2.31.0
rich>=13.7.0
